requirements.txt
Django==3.1.7
django-ckeditor==6.0.0
six==1.15.0
django-uuslug==1.2.0
Pillow==8.1.2